\name{snpXpert}
\alias{snpXpert}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Software for analyzing SNP data}
\description{
snpXpert is a software to analyze SNP data.
This software provides its own GUI to help biologists who are not familiar with R commands.
}
\usage{
snpXpert()
}


\author{ Eun-kyung Lee, Dankyu Yoon, Taesung Park}
\examples{
snpXpert()
}
\keyword{ dplot }% at least one, from doc/KEYWORDS
\keyword{ utilities }% __ONLY ONE__ keyword per line
